Update SDK to version 25.06.0 for iMXRT10xx

+++ middleware/nxp_iot_agent/src/common/nxp_iot_agent_utils_protobuf.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2024 NXP
+ * Copyright 2018-2025 NXP
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -13,8 +13,8 @@ static bool copy_bytes_to_buffer(pb_istream_t *stream, buffer_t *read_buffer)
 {
     size_t len = stream->bytes_left;
     if (len > read_buffer->len) {
-        IOT_AGENT_ERROR("received byte stream is too long [%d bytes] to "
-            "fit in buffer with size [%d bytes]", (int)len, (int)read_buffer->len);
+        IOT_AGENT_ERROR("received byte stream is too long [%zu bytes] to "
+            "fit in buffer with size [%zu bytes]", len, read_buffer->len);
         return false;
     }
     if (!pb_read(stream, (pb_byte_t*)(read_buffer->buf), len)) {
